feat: migrate optimize-js lowering to HIR with legacy parity
- add HIR-root entry points and compile_lowered for precomputed hir_js results
- rework s2i lowering to consume HIR bodies while keeping legacy AST path behind feature
- map semantic-js bindings by span to preserve symbol classification and eval handling
- fix hir-js function param lowering and reject TS-only export-as-namespace in emit-js

+++ optimize-js/src/il/inst.rs
@@ -175,7 +175,7 @@ fn is_dummy_symbol(sym: &SymbolId) -> bool {
 }
 
 fn dummy_symbol() -> SymbolId {
-  SymbolId(u32::MAX)
+  SymbolId(u32::MAX as u64)
 }
 
 #[derive(PartialEq, Eq, Clone, Debug, Serialize)]
